backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
Django==5.0.9
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
